+++ backend/src/modules/user/entities/user.entity.ts
@@ -1,0 +1,13 @@
+import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity('user')
+export class UserEntity {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column({ type: 'varchar', length: 150, unique: true, nullable: false })
+  name: string;
+
+  @Column({ type: 'timestamp', nullable: false, default: () => 'now()' })
+  created_at: Date;
+}