+++ backend/src/modules/user/user.module.ts
@@ -1,0 +1,17 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { UserController } from './user.controller';
+import { UserEntity } from './entities/user.entity';
+import { UserService } from './user.service';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([UserEntity])],
+  controllers: [UserController],
+  providers: [
+    UserService,
+    { provide: 'USER_SERVICE', useClass: UserService },
+  ],
+  exports: ['USER_SERVICE', TypeOrmModule],
+})
+export class UserModule {}